refactor(ai): rename gemini module to ai module
