+++ apps/api/src/modules/ai/gemini.module.ts
