+++ apps/api/src/modules/ai/gemini.service.ts
